+++ cli/generate.go
@@ -47,7 +47,8 @@ var (
 )
 
 const (
-	execution, consensus, validator, mevBoost = "execution", "consensus", "validator", "mev-boost"
+	execution, consensus, validator, mevBoost, optimism = "execution", "consensus", "validator", "mev-boost", "optimism"
+	jwtPathName                                         = "jwtsecret"
 )
 
 type CustomFlags struct {
@@ -57,9 +58,18 @@ type CustomFlags struct {
 	customDeployBlock   string
 }
 
+type OptimismFlags struct {
+	optimismName          string
+	optimismExecutionName string
+	elOpExtraFlags        []string
+	opExtraFlags          []string
+	isBase                bool
+}
+
 // GenCmdFlags is a struct that holds the flags of the generate command
 type GenCmdFlags struct {
 	CustomFlags
+	OptimismFlags
 	executionName     string
 	consensusName     string
 	validatorName     string
@@ -98,6 +108,7 @@ It will create a 'docker-compose.yml' and a '.env', which you will need later to
 You can generate:
 - Full Node (execution + consensus + validator)
 - Full Node without Validator (execution + consensus)
+- Optimism Full Node
 - Execution Node
 - Consensus Node
 - Validator Node
@@ -112,6 +123,7 @@ You can generate:
 	cmd.AddCommand(ConsensusSubCmd(sedgeAction))
 	cmd.AddCommand(ValidatorSubCmd(sedgeAction))
 	cmd.AddCommand(MevBoostSubCmd(sedgeAction))
+	cmd.AddCommand(OpFullNodeSubCmd(sedgeAction))
 
 	cmd.PersistentFlags().BoolVar(&lidoNode, "lido", false, "generate Lido CSM node")
 	cmd.PersistentFlags().StringVarP(&generationPath, "path", "p", configs.DefaultAbsSedgeDataPath, "generation path for sedge data. Default is sedge-data")
@@ -250,12 +262,20 @@ func runGenCmd(out io.Writer, flags *GenCmdFlags, sedgeAction actions.SedgeActio
 
 	// Generate jwt secrets if needed
 	if flags.jwtPath == "" {
-		flags.jwtPath, err = handleJWTSecret(generationPath)
+		flags.jwtPath, err = handleJWTSecret(generationPath, jwtPathName)
 		if err != nil {
 			return err
 		}
 	} else {
 		flags.jwtPath, err = loadJWTSecret(flags.jwtPath)
+		if err != nil {
+			return err
+		}
+	}
+	var jwtSecretOP string
+	// If optimism is included in the services, generate the jwt secret for it
+	if utils.Contains(services, optimism) {
+		jwtSecretOP, err = handleJWTSecret(generationPath, jwtPathName+"-op")
 		if err != nil {
 			return err
 		}
@@ -275,11 +295,25 @@ func runGenCmd(out io.Writer, flags *GenCmdFlags, sedgeAction actions.SedgeActio
 
 	vlStartGracePeriod := configs.NetworkEpochTime(network) * time.Duration(flags.waitEpoch)
 
+	var consensusApiUrl string
+	var executionApiUrl string
+	var executionAuthUrl string
+	if combinedClients.Consensus == nil {
+		consensusApiUrl = flags.consensusApiUrl
+	}
+
+	if combinedClients.Execution == nil {
+		executionApiUrl = flags.executionApiUrl
+		executionAuthUrl = flags.executionAuthUrl
+	}
+
 	// Generate docker-compose scripts
 	gd := generate.GenData{
 		ExecutionClient:         combinedClients.Execution,
 		ConsensusClient:         combinedClients.Consensus,
 		ValidatorClient:         combinedClients.Validator,
+		ExecutionOPClient:       combinedClients.ExecutionOP,
+		OptimismClient:          combinedClients.Optimism,
 		Network:                 network,
 		CheckpointSyncUrl:       flags.checkpointSyncUrl,
 		FeeRecipient:            flags.feeRecipient,
@@ -289,6 +323,9 @@ func runGenCmd(out io.Writer, flags *GenCmdFlags, sedgeAction actions.SedgeActio
 		ElExtraFlags:            flags.elExtraFlags,
 		ClExtraFlags:            flags.clExtraFlags,
 		VlExtraFlags:            flags.vlExtraFlags,
+		ElOpExtraFlags:          flags.elOpExtraFlags,
+		OpExtraFlags:            flags.opExtraFlags,
+		IsBase:                  flags.isBase,
 		MapAllPorts:             flags.mapAllPorts,
 		Mev:                     !flags.noMev && utils.Contains(services, validator) && utils.Contains(services, consensus) && !flags.noValidator,
 		MevImage:                flags.mevImage,
@@ -298,9 +335,9 @@ func runGenCmd(out io.Writer, flags *GenCmdFlags, sedgeAction actions.SedgeActio
 		MevBoostEndpoint:        flags.mevBoostUrl,
 		Services:                services,
 		VLStartGracePeriod:      uint(vlStartGracePeriod.Seconds()),
-		ExecutionApiUrl:         flags.executionApiUrl,
-		ExecutionAuthUrl:        flags.executionAuthUrl,
-		ConsensusApiUrl:         flags.consensusApiUrl,
+		ExecutionApiUrl:         executionApiUrl,
+		ExecutionAuthUrl:        executionAuthUrl,
+		ConsensusApiUrl:         consensusApiUrl,
 		ECBootnodes:             flags.customEnodes,
 		CCBootnodes:             flags.customEnrs,
 		CustomChainSpecPath:     flags.CustomFlags.customChainSpec,
@@ -311,6 +348,7 @@ func runGenCmd(out io.Writer, flags *GenCmdFlags, sedgeAction actions.SedgeActio
 		MevBoostOnValidator:     flags.mevBoostOnVal,
 		ContainerTag:            containerTag,
 		LatestVersion:           flags.latestVersion,
+		JWTSecretOP:             jwtSecretOP,
 	}
 	_, err = sedgeAction.Generate(actions.GenerateOptions{
 		GenerationData: gd,
@@ -338,7 +376,7 @@ func runGenCmd(out io.Writer, flags *GenCmdFlags, sedgeAction actions.SedgeActio
 }
 
 func valClients(allClients clients.OrderedClients, flags *GenCmdFlags, services []string) (*clients.Clients, error) {
-	var executionClient, consensusClient, validatorClient *clients.Client
+	var executionClient, consensusClient, validatorClient, executionOpClient, opClient *clients.Client
 	var err error
 
 	// execution client
@@ -353,6 +391,7 @@ func valClients(allClients clients.OrderedClients, flags *GenCmdFlags, services 
 			if len(executionParts) > 1 {
 				log.Warn(configs.CustomExecutionImagesWarning)
 				executionClient.Image = strings.Join(executionParts[1:], ":")
+				flags.latestVersion = false
 			}
 		}
 		executionClient.SetImageOrDefault(strings.Join(executionParts[1:], ":"))
@@ -374,6 +413,7 @@ func valClients(allClients clients.OrderedClients, flags *GenCmdFlags, services 
 			if len(consensusParts) > 1 {
 				log.Warn(configs.CustomConsensusImagesWarning)
 				consensusClient.Image = strings.Join(consensusParts[1:], ":")
+				flags.latestVersion = false
 			}
 		}
 		consensusClient.SetImageOrDefault(strings.Join(consensusParts[1:], ":"))
@@ -395,6 +435,7 @@ func valClients(allClients clients.OrderedClients, flags *GenCmdFlags, services 
 			if len(validatorParts) > 1 {
 				log.Warn(configs.CustomValidatorImagesWarning)
 				validatorClient.Image = strings.Join(validatorParts[1:], ":")
+				flags.latestVersion = false
 
 			}
 		}
@@ -405,11 +446,53 @@ func valClients(allClients clients.OrderedClients, flags *GenCmdFlags, services 
 	} else {
 		validatorClient = nil
 	}
+	// optimism client
+	if utils.Contains(services, optimism) {
+		optimismParts := strings.Split(flags.optimismName, ":")
+		opClient, err = clients.RandomChoice(allClients[optimism])
+		if err != nil {
+			return nil, err
+		}
+		if flags.optimismName != "" {
+			opClient.Name = "optimism"
+			if len(optimismParts) > 1 {
+				opClient.Image = strings.Join(optimismParts[1:], ":")
+			}
+		}
+		opClient.SetImageOrDefault(strings.Join(optimismParts[1:], ":"))
+		if err = clients.ValidateClient(opClient, optimism); err != nil {
+			return nil, err
+		}
+
+		optimismExecutionParts := strings.Split(flags.optimismExecutionName, ":")
+		executionOpClient = allClients[execution]["nethermind"]
+		if flags.optimismExecutionName != "" {
+			executionOpClient.Name = optimismExecutionParts[0]
+			if len(optimismExecutionParts) > 1 {
+				executionOpClient.Image = strings.Join(optimismExecutionParts[1:], ":")
+			}
+		}
+		executionOpClient.SetImageOrDefault(strings.Join(optimismExecutionParts[1:], ":"))
+		if err = clients.ValidateClient(executionOpClient, optimism); err != nil {
+			return nil, err
+		}
+
+		// If set execution-api-url, set execution and beacon to nil
+		if flags.executionApiUrl != "" {
+			executionClient = nil
+			consensusClient = nil
+		}
+	} else {
+		opClient = nil
+		executionOpClient = nil
+	}
 
 	return &clients.Clients{
-		Execution: executionClient,
-		Consensus: consensusClient,
-		Validator: validatorClient,
+		Execution:   executionClient,
+		Consensus:   consensusClient,
+		Validator:   validatorClient,
+		ExecutionOP: executionOpClient,
+		Optimism:    opClient,
 	}, err
 }
 
@@ -434,7 +517,7 @@ func initGenPath(path string) error {
 	return nil
 }
 
-func handleJWTSecret(generationPath string) (string, error) {
+func handleJWTSecret(generationPath, name string) (string, error) {
 	log.Info(configs.GeneratingJWTSecret)
 	if !filepath.IsAbs(generationPath) {
 		return "", fmt.Errorf(configs.GenerateJWTSecretError, fmt.Errorf("generation path must be absolute"))
@@ -445,7 +528,7 @@ func handleJWTSecret(generationPath string) (string, error) {
 		return "", fmt.Errorf(configs.GenerateJWTSecretError, err)
 	}
 
-	jwtPath, err := filepath.Abs(filepath.Join(generationPath, "jwtsecret"))
+	jwtPath, err := filepath.Abs(filepath.Join(generationPath, name))
 	if err != nil {
 		return "", fmt.Errorf(configs.GenerateJWTSecretError, err)
 	}