+++ CHANGELOG.md
@@ -13,6 +13,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 -  Teku and Lighthouse import keys container error on Windows.
 
+## [v1.5.0] - 2024-09-06
+
+### Added
+- Add support for Optimism and Base, using Nethermind Client on Mainnet and Sepolia.
+
+### Changed
+- Update client images to latest versions.
+
+#### Fixed
+- Remove Peer upper limit of peers on CL
+
 ##  [v1.4.0] - 2024-07-10
 
 ### Fixed
@@ -41,7 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Update client images to Dencun-ready versions. 
+- Update client images to Dencun-ready versions.
 
 ## [v1.3.1] - 2024-02-14